chore: move wallets react code to a subfolder called legacy
this is part of preparation for our upcomin wallet management
