+++ wallets/react/src/index.ts
@@ -1,7 +1,7 @@
-export * from './helpers';
+export * from './legacy/helpers';
 export { default as Provider } from './provider';
-export { useWallets } from './hooks';
-export * from './types';
+export { useWallets } from './legacy/hooks';
+export * from './legacy/types';
 
 export type { EventHandler } from '@rango-dev/wallets-core';
 export { Events, readAccountAddress } from '@rango-dev/wallets-core';
