+++ wallets/react/src/legacy/constants.ts
