+++ wallets/react/src/legacy/context.ts
