+++ wallets/react/src/legacy/hooks.ts
@@ -1,9 +1,10 @@
+import type { ProviderContext, WalletActions, WalletConfig } from './types';
+import type { EventHandler as WalletEventHandler } from '@rango-dev/wallets-core';
+
+import Wallet from '@rango-dev/wallets-core';
 import { useContext, useRef } from 'react';
+
 import { WalletContext } from './context';
-import { ProviderContext, WalletActions, WalletConfig } from './types';
-import Wallet, {
-  EventHandler as WalletEventHandler,
-} from '@rango-dev/wallets-core';
 
 /**
  *
@@ -14,9 +15,11 @@ export function useInitializers(onChangeState: WalletEventHandler) {
     [key: string]: Wallet | undefined;
   }>({});
 
-  // If `wallet` hasn't been added to `availableWallets`,
-  // Get a instance of `Wallet` and save the refrence in `availableWallets`.
-  // Otherwise, return the already created instance.
+  /*
+   * If `wallet` hasn't been added to `availableWallets`,
+   * Get a instance of `Wallet` and save the refrence in `availableWallets`.
+   * Otherwise, return the already created instance.
+   */
   function updater(wallet: {
     actions: WalletActions;
     config: WalletConfig;
@@ -41,7 +44,8 @@ export function useInitializers(onChangeState: WalletEventHandler) {
 
 export function useWallets(): ProviderContext {
   const context = useContext(WalletContext);
-  if (!context)
+  if (!context) {
     throw Error('useWallet can only be used within the Provider component');
+  }
   return context;
 }
