+++ wallets/react/src/legacy/types.ts
